pyproject.toml
[project]
name = "aith-book-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Maxim Abramov",email = "abramovmaxim.894@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "pydantic (>=2.12.4,<3.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-openai (>=1.0.3,<2.0.0)",
    "langchain-ollama (>=1.0.0,<2.0.0)",
    "neo4j (>=6.0.3,<7.0.0)",
    "ruff (>=0.14.8,<0.15.0)",
    "pre-commit (>=4.5.0,<5.0.0)",
    "gdown (>=5.2.0,<6.0.0)",
    "fastapi (>=0.124.2,<0.125.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "mistralai (>=1.10.0,<2.0.0)",
    "ragas[langchain] (>=0.4.2,<0.5.0)",
    "datasets (>=4.4.2,<5.0.0)",
    "langchain-mistralai (>=1.1.1,<2.0.0)",
    "langchain-deepseek (>=1.0.1,<2.0.0)",
    "openai (>=2.14.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.format]
quote-style = "double"
exclude = ["data/**"]
indent-style = "space"
skip-magic-trailing-comma = true